@@ -2,11 +2,11 @@ apply plugin: 'com.android.library'
 // apply from: 'maven-push.gradle'
 
 ext {
-    supportLibVersion = '25.3.1'
+    supportLibVersion = '25.4.0'
 }
 
 android {
-    compileSdkVersion 25
+    compileSdkVersion 26
     buildToolsVersion '25.0.3'
 
     dataBinding {
@@ -14,7 +14,7 @@ android {
     }
     defaultConfig {
         minSdkVersion 14
-        targetSdkVersion 25
+        targetSdkVersion 26
         versionCode 42
         versionName "0.14.0"
         renderscriptTargetApi 20
@@ -33,14 +33,14 @@ android {
 }
 
 dependencies {
-    compile fileTree(include: ['*.jar'], dir: 'libs')
-    compile "com.android.support:appcompat-v7:$supportLibVersion"
-    compile "com.android.support:support-v4:$supportLibVersion"
-    compile "com.android.support:gridlayout-v7:$supportLibVersion"
-    compile "com.android.support:recyclerview-v7:$supportLibVersion"
-    compile "com.android.support:appcompat-v7:$supportLibVersion"
-    compile "com.android.support:design:$supportLibVersion"
-    compile 'de.greenrobot:androidsvg:1.2.2-beta-1-tweaked-2'
-    compile 'com.android.support.constraint:constraint-layout:1.0.2'
-    compile 'com.annimon:stream:1.1.6'
+    annotationProcessor 'com.android.databinding:compiler:3.0.0-alpha3'
+    compile("com.android.support:appcompat-v7:$supportLibVersion") {
+        force = true
+    }
+    implementation "com.android.support:gridlayout-v7:$supportLibVersion"
+    implementation "com.android.support:recyclerview-v7:$supportLibVersion"
+    implementation "com.android.support:design:$supportLibVersion"
+    implementation 'de.greenrobot:androidsvg:1.2.2-beta-1-tweaked-2'
+    implementation 'com.android.support.constraint:constraint-layout:1.0.2'
+    implementation 'com.annimon:stream:1.1.7'
 }
