@@ -1,7 +1,11 @@
 apply plugin: 'com.android.application'
 
+ext {
+    supportLibVersion = '25.4.0'
+}
+
 android {
-    compileSdkVersion 25
+    compileSdkVersion 26
     buildToolsVersion '25.0.3'
 
     dataBinding {
@@ -10,7 +14,7 @@ android {
     defaultConfig {
         applicationId "tk.zielony.carbonsamples"
         minSdkVersion 14
-        targetSdkVersion 25
+        targetSdkVersion 26
         versionCode 1
         versionName "1.0"
         renderscriptTargetApi 20
@@ -30,11 +34,16 @@ android {
 }
 
 dependencies {
-    compile project(':carbon')
-    compile 'com.squareup.picasso:picasso:2.4.0'
-    compile 'com.annimon:stream:1.1.4'
-    compile 'com.github.ZieIony.RandomData:person:812a7c222a'
-    compile 'com.github.ZieIony:LandscapeView:94187526bb'
-    compile 'com.android.support:cardview-v7:25.3.1'
-    compile project(':rxcarbon')
+    annotationProcessor 'com.android.databinding:compiler:3.0.0-alpha3'
+    implementation project(':carbon')
+    implementation project(':rxcarbon')
+    implementation 'com.squareup.picasso:picasso:2.5.2'
+    implementation 'com.annimon:stream:1.1.7'
+    implementation 'com.jakewharton.rxbinding2:rxbinding:2.0.0'
+    implementation "com.android.support:recyclerview-v7:$supportLibVersion"
+    implementation "com.android.support:cardview-v7:$supportLibVersion"
+    implementation "com.android.support:design:$supportLibVersion"
+    implementation 'com.github.ZieIony:randomdata:05c14279e1'
+    implementation 'com.github.ZieIony.RandomData:person:05c14279e1'
+    implementation 'com.github.ZieIony:LandscapeView:94187526bb'
 }
