@@ -63,7 +63,7 @@ import carbon.view.StrokeView;
 import carbon.view.VisibleView;
 import carbon.widget.OnInsetsChangedListener;
 import carbon.widget.RenderingMode;
-import carbon.widget.TouchMarginView;
+import carbon.view.TouchMarginView;
 
 public class AppBarLayout extends android.support.design.widget.AppBarLayout
         implements
