@@ -72,6 +72,8 @@ import carbon.shadow.ShadowView;
 import carbon.view.AutoSizeTextView;
 import carbon.view.RevealView;
 import carbon.view.StateAnimatorView;
+import carbon.view.TintedView;
+import carbon.view.TouchMarginView;
 import carbon.view.ValidStateView;
 import carbon.view.VisibleView;
 
@@ -1299,8 +1301,12 @@ public class EditText extends android.widget.EditText
         return maxTextSize;
     }
 
-    public float getAutoSizeStepGranularity() {
-        return autoSizeStepGranularity;
+    public int getAutoSizeStepGranularity() {
+        return (int) autoSizeStepGranularity;
+    }
+
+    public void setAutoSizeStepGranularity(int autoSizeStepGranularity) {
+        setAutoSizeStepGranularity((float) autoSizeStepGranularity);
     }
 
     public void setAutoSizeStepGranularity(float autoSizeStepGranularity) {
